% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_update.R
\name{generate_checksum_input}
\alias{generate_checksum_input}
\title{User Written Function to Create Input for Checksum Comparison}
\usage{
generate_checksum_input()
}
\arguments{
\item{key}{API key for the KOF Time Series Database.}
}
\description{
This function generates input for computation of checksums to identify
outdated content. Good inputs are either publication dates extracted from
official publisher sites or APIs or any single time series from a database,
because opentsi definition all time series of the same dataset must
have the same publication date.
}
